# swagger.yaml
openapi: 3.0.3
info:
  title: ARMA Audit Engine API
  description: |
    GrowthScope / ARMA Audit Engine — generates local SEO audit reports for small businesses.

    **Primary endpoint (used by the dashboard UI):**
    - `POST /arma-report` — self-contained 5-page briefing vs a **manually supplied competitor**
      (`competitor_url`, optional `competitor_position`). One call, no prerequisites.

    **Legacy workflow (API-only, no longer in the UI):**
    1. `POST /lite-report` — scores a business, locks competitor data, returns a PDF (or JSON) lite report
    2. `POST /full-report` — deep-dive audit PDF; reads the competitor locked by step 1

    **Required env vars:** `DATAFORSEO_LOGIN` + `DATAFORSEO_PASSWORD` (primary map pack source),
    `GOOGLE_PLACES_API_KEY` (fallback), `ANTHROPIC_API_KEY` (analysis), `PAGESPEED_API_KEY` (scores).
  version: 1.0.0
  contact:
    email: pankajkumar.techie@gmail.com
servers:
  - url: https://www.arma-agency.site
    description: Production server
  - url: http://localhost:3002
    description: Local development server
...
paths:
...
  /mappack-debug:
    get:
      tags: [Debug]
      summary: Inspect Map Pack Rankings
      description: |
        Returns Google Maps pack positions for a vertical + city, showing exactly what
        the audit engine uses in reports.

        **Source priority:**
        1. SQLite cache (populated during real audit runs — exact data used in the last report)
        2. DataForSEO SERP Maps live (if `DATAFORSEO_LOGIN` env var is set)
        3. Google Places API v1 text search (approximate prominence order, not guaranteed Map Pack rank)

        Pass `bust=1` to clear the cache for this keyword/city/state and force a fresh fetch.
      parameters:
        - in: query
          name: vertical
          required: true
          schema:
            type: string
          description: Business vertical / service category
          example: Roofing
        - in: query
          name: city
          required: true
          schema:
            type: string
          description: City name
          example: Anchorage
        - in: query
          name: state
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 2
          description: Two-letter US state abbreviation
          example: AK
        - in: query
          name: bust
          schema:
            type: string
            enum: ['1']
          description: Pass `1` to clear the cache entry and force a live fetch
      responses:
        '200':
          description: Map pack positions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapPackDebugResponse'
              example:
                vertical: Roofing
                city: Anchorage
                state: AK
                source: cache (fetched 2026-06-08 07:00:00)
                note: Cached data from the audit engine — these ranks are exactly what was used in the last report.
                count: 5
                positions:
                  - rank: 1
                    name: ABC Roofing
                    rating: 4.8
                    reviews: 312
                    place_id: ChIJXXXXXXXXXX
                    maps_url: https://www.google.com/maps/place/?q=place_id:ChIJXXXXXXXXXX
        '400':
          description: Missing required query params
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "vertical, city, state required"
        '500':
          description: No data sources configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /pagespeed-check:
    get:
      tags: [Debug]
      summary: Check PageSpeed Scores
      description: |
        Runs Google PageSpeed Insights for the given URL on **both** mobile and desktop,
        regardless of the `strategy` param. Results are cached in SQLite; pass `bust=1`
        to force a fresh API call.
      parameters:
        - in: query
          name: url
          required: true
          schema:
            type: string
            format: uri
          description: Full URL to test
          example: https://example.com
        - in: query
          name: strategy
          schema:
            type: string
            enum: [mobile, desktop]
            default: mobile
          description: Hint only — both strategies are always returned
        - in: query
          name: bust
          schema:
            type: string
            enum: ['1', 'true']
          description: Pass `1` or `true` to bypass the SQLite cache
      responses:
        '200':
          description: PageSpeed scores for mobile and desktop
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageSpeedResponse'
        '400':
          description: Missing URL or invalid strategy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── System ───────────────────────────────────────────────────────────────────
...
  /cache-clear:
    delete:
      tags: [Cache]
      summary: Clear Map Pack Cache
      description: |
        Deletes map pack cache entries.
        - Provide `keyword + city + state` to clear a single entry.
        - Omit all params to clear **every** cache entry.
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
          description: Vertical keyword (e.g. `Roofing`)
        - in: query
          name: city
          schema:
            type: string
          description: City name
        - in: query
          name: state
          schema:
            type: string
          description: State abbreviation
      responses:
        '200':
          description: Number of cleared entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheClearResponse'
              examples:
                single:
                  value:
                    cleared: 1
                    entry: "Roofing / Anchorage, AK"
                all:
                  value:
                    cleared: 14
                    message: "All map pack cache entries deleted"

# ── Components ────────────────────────────────────────────────────────────────
components:
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    MapPackPosition:
      type: object
      properties:
        rank:
          type: integer
          description: Position in the map pack (1 = top result)
        name:
          type: string
        rating:
          type: number
          format: float
        reviews:
          type: integer
        place_id:
          type: string
          nullable: true
        maps_url:
          type: string
          format: uri
          nullable: true
        data_source:
          type: string
          description: Present only in cache responses (`dataforseo_maps` or `google_places`)
    MapPackDebugResponse:
      type: object
      properties:
        vertical:
          type: string
        city:
          type: string
        state:
          type: string
        source:
          type: string
          description: |
            Where the data came from:
            - `cache (fetched <timestamp>)` — SQLite cache from a real audit run
            - `dataforseo_maps_live` — live DataForSEO SERP Maps API
            - `google_places_v1_live` — live Google Places text search (approximate)
        note:
          type: string
        count:
          type: integer
        positions:
          type: array
          items:
            $ref: '#/components/schemas/MapPackPosition'

    # ── PageSpeed ────────────────────────────────────────────────────────────────
    PageSpeedResult:
      type: object
      properties:
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: Performance score
        lcp:
          type: string
          description: Largest Contentful Paint (e.g. `"2.5 s"`)
        cls:
          type: string
          description: Cumulative Layout Shift
        is_fallback:
          type: boolean
          description: True if the score could not be fetched and a fallback value was used
        cached:
          type: boolean
          description: True if served from SQLite cache
    PageSpeedResponse:
      type: object
      properties:
        url:
          type: string
        tested_at:
          type: string
          format: date-time
        mobile:
          $ref: '#/components/schemas/PageSpeedResult'
        desktop:
          $ref: '#/components/schemas/PageSpeedResult'
        note:
          type: string

    # ── Health / Cache ───────────────────────────────────────────────────────────
...
    CacheClearResponse:
      type: object
      properties:
        cleared:
          type: integer
          description: Number of SQLite rows deleted
        message:
          type: string
          description: Present when all entries were cleared
        entry:
          type: string
          description: Present when a specific `keyword / city, state` entry was cleared

    # ── Lite report JSON response ────────────────────────────────────────────────
...